requirements.txt
numpy
scipy
matplotlib
tqdm
Pillow
streamlit
